backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
PyMuPDF==1.25.3
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
chromadb>=0.6.0
google-genai>=1.0.0
youtube-transcript-api>=0.6.3
yt-dlp>=2024.0.0
python-dotenv>=1.0.0
